do not return None from harvest
It can happen, that harvest returns None, e.g. when an endpoint returns
503 Service Unavailable repeatedly.

However, there is an assumption that response will not be None during
iteration:

    self.oai_response = self.sickle.harvest(**params)
    error = self.oai_response.xml.find(

This yields an 'NoneType' object has no attribute 'xml' error.

Hence, when max_retries is exceeded, a TimeoutError would be a bit more
clear.