Fix 'NoneType' object has no attribute 'getchildren' error.
We're running into an issue where we are trying to harvest an OAI
endpoint where one of the records has no metadata and its status is not
set to deleted.

The error we get for this record is at iteration time:

```
AttributeError: 'NoneType' object has no attribute 'getchildren'
```

This change updates the code so that it skips a record that has no
metadata. Thus, the exception is not thrown.